Google Authentication Manager for Flask
This module initializes the cookies, filesystem, and prepares oauth. The class itself handles authentication with Google. Particularly, loading important information Google needs, sending the user to Google, and handling when the user returns from Google. When the user returns the method `get_callback_token`, it returns the token JSON object back to the main program.